PRE-3469: Promote SyliusTokenCache to PayplugTokenCache
Moves the ITokenCache skeleton out of src/Spike/ into a real production
namespace. Stays validated by a real-infrastructure integration test only
(see SpikeIntegrationTest) — no production call site, since its one real
analog (PayPlugApiClientFactory's OAuth token cache) gates authentication
for every gateway and isn't worth the regression risk of replacing here.
